Cache client marshallers per schema
