+++ modules/client/src/main/java/org/apache/ignite/internal/client/table/ClientRecordView.java
@@ -1,38 +1,15 @@
 package org.apache.ignite.internal.client.table;
 
-import static org.apache.ignite.internal.client.proto.ClientDataType.BIGINTEGER;
-import static org.apache.ignite.internal.client.proto.ClientDataType.BITMASK;
-import static org.apache.ignite.internal.client.proto.ClientDataType.BOOLEAN;
-import static org.apache.ignite.internal.client.proto.ClientDataType.BYTES;
-import static org.apache.ignite.internal.client.proto.ClientDataType.DATE;
-import static org.apache.ignite.internal.client.proto.ClientDataType.DATETIME;
-import static org.apache.ignite.internal.client.proto.ClientDataType.DECIMAL;
-import static org.apache.ignite.internal.client.proto.ClientDataType.DOUBLE;
-import static org.apache.ignite.internal.client.proto.ClientDataType.FLOAT;
-import static org.apache.ignite.internal.client.proto.ClientDataType.INT16;
-import static org.apache.ignite.internal.client.proto.ClientDataType.INT32;
-import static org.apache.ignite.internal.client.proto.ClientDataType.INT64;
-import static org.apache.ignite.internal.client.proto.ClientDataType.INT8;
-import static org.apache.ignite.internal.client.proto.ClientDataType.NUMBER;
-import static org.apache.ignite.internal.client.proto.ClientDataType.STRING;
-import static org.apache.ignite.internal.client.proto.ClientDataType.TIME;
-import static org.apache.ignite.internal.client.proto.ClientDataType.TIMESTAMP;
-
 import java.io.Serializable;
 import java.util.Collection;
 import java.util.Map;
 import java.util.Objects;
 import java.util.concurrent.CompletableFuture;
-import org.apache.ignite.internal.client.proto.ClientDataType;
 import org.apache.ignite.internal.client.proto.ClientOp;
-import org.apache.ignite.internal.client.proto.TuplePart;
-import org.apache.ignite.internal.marshaller.BinaryMode;
 import org.apache.ignite.internal.marshaller.ClientMarshallerReader;
 import org.apache.ignite.internal.marshaller.ClientMarshallerWriter;
 import org.apache.ignite.internal.marshaller.Marshaller;
-import org.apache.ignite.internal.marshaller.MarshallerColumn;
 import org.apache.ignite.internal.marshaller.MarshallerException;
-import org.apache.ignite.lang.IgniteException;
 import org.apache.ignite.table.InvokeProcessor;
 import org.apache.ignite.table.RecordView;
 import org.apache.ignite.table.mapper.Mapper;
@@ -73,8 +50,7 @@ public class ClientRecordView<R> implements RecordView<R> {
                     out.packInt(schema.version());
 
                     try {
-                        // TODO: Cache marshallers per schema.
-                        Marshaller keyMarsh = getMarshaller(schema, TuplePart.KEY);
+                        Marshaller keyMarsh = schema.getKeyMarshaller(recMapper);
 
                         keyMarsh.writeObject(keyRec, new ClientMarshallerWriter(out));
                     } catch (MarshallerException e) {
@@ -83,13 +59,13 @@ public class ClientRecordView<R> implements RecordView<R> {
                     }
                 },
                 (inSchema, in) -> {
-                    var marsh = getMarshaller(inSchema, TuplePart.VAL);
+                    var marsh = inSchema.getValMarshaller(recMapper);
 
                     try {
                         var res = (R) marsh.readObject(new ClientMarshallerReader(in), null);
 
                         if (!marsh.isSimple()) {
-                            Marshaller keyMarsh = getMarshaller(inSchema, TuplePart.KEY);
+                            Marshaller keyMarsh = inSchema.getKeyMarshaller(recMapper);
 
                             keyMarsh.copyObject(keyRec, res);
                         }
@@ -271,91 +247,5 @@ public class ClientRecordView<R> implements RecordView<R> {
     @Override
     public RecordView<R> withTransaction(Transaction tx) {
         return null;
-    }
-
-    private Marshaller getMarshaller(ClientSchema schema, TuplePart part) {
-        ClientColumn[] schemaCols = schema.columns();
-
-        int colCount = schemaCols.length;
-        int firstColIdx = 0;
-
-        if (part == TuplePart.KEY) {
-            colCount = schema.keyColumnCount();
-        } else if (part == TuplePart.VAL) {
-            colCount = schemaCols.length - schema.keyColumnCount();
-            firstColIdx = schema.keyColumnCount();
-        }
-
-        MarshallerColumn[] cols = new MarshallerColumn[colCount];
-
-        for (int i = 0; i < colCount; i++) {
-            var col = schemaCols[i  + firstColIdx];
-
-            cols[i] = new MarshallerColumn(col.name(), mode(col.type()));
-        }
-
-        return Marshaller.createMarshaller(cols, recMapper, part == TuplePart.KEY);
-    }
-
-    private static BinaryMode mode(int dataType) {
-        switch (dataType) {
-            case BOOLEAN:
-                throw new IgniteException("TODO: " + dataType);
-
-            case INT8:
-                // TODO: P_BYTE?
-                return BinaryMode.BYTE;
-
-            case INT16:
-                return BinaryMode.SHORT;
-
-            case INT32:
-                return BinaryMode.INT;
-
-            case INT64:
-                return BinaryMode.LONG;
-
-            case FLOAT:
-                return BinaryMode.FLOAT;
-
-            case DOUBLE:
-                return BinaryMode.DOUBLE;
-
-            case ClientDataType.UUID:
-                return BinaryMode.UUID;
-
-            case STRING:
-                return BinaryMode.STRING;
-
-            case BYTES:
-                return BinaryMode.BYTE_ARR;
-
-            case DECIMAL:
-                return BinaryMode.DECIMAL;
-
-            case BIGINTEGER:
-                return BinaryMode.NUMBER;
-
-            case BITMASK:
-                return BinaryMode.BITSET;
-
-            case NUMBER:
-                return BinaryMode.NUMBER;
-
-            case DATE:
-                return BinaryMode.DATE;
-
-            case TIME:
-                return BinaryMode.TIME;
-
-            case DATETIME:
-                return BinaryMode.DATETIME;
-
-            case TIMESTAMP:
-                return BinaryMode.TIMESTAMP;
-
-            default:
-                throw new IgniteException("Unknown client data type: " + dataType);
-        }
     }
 }
