+++ modules/client/src/main/java/org/apache/ignite/internal/client/table/ClientSchema.java
@@ -17,15 +17,41 @@
 
 package org.apache.ignite.internal.client.table;
 
+import static org.apache.ignite.internal.client.proto.ClientDataType.BIGINTEGER;
+import static org.apache.ignite.internal.client.proto.ClientDataType.BITMASK;
+import static org.apache.ignite.internal.client.proto.ClientDataType.BOOLEAN;
+import static org.apache.ignite.internal.client.proto.ClientDataType.BYTES;
+import static org.apache.ignite.internal.client.proto.ClientDataType.DATE;
+import static org.apache.ignite.internal.client.proto.ClientDataType.DATETIME;
+import static org.apache.ignite.internal.client.proto.ClientDataType.DECIMAL;
+import static org.apache.ignite.internal.client.proto.ClientDataType.DOUBLE;
+import static org.apache.ignite.internal.client.proto.ClientDataType.FLOAT;
+import static org.apache.ignite.internal.client.proto.ClientDataType.INT16;
+import static org.apache.ignite.internal.client.proto.ClientDataType.INT32;
+import static org.apache.ignite.internal.client.proto.ClientDataType.INT64;
+import static org.apache.ignite.internal.client.proto.ClientDataType.INT8;
+import static org.apache.ignite.internal.client.proto.ClientDataType.NUMBER;
+import static org.apache.ignite.internal.client.proto.ClientDataType.STRING;
+import static org.apache.ignite.internal.client.proto.ClientDataType.TIME;
+import static org.apache.ignite.internal.client.proto.ClientDataType.TIMESTAMP;
+
 import java.util.HashMap;
 import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
+import org.apache.ignite.internal.client.proto.ClientDataType;
+import org.apache.ignite.internal.client.proto.TuplePart;
+import org.apache.ignite.internal.marshaller.BinaryMode;
+import org.apache.ignite.internal.marshaller.Marshaller;
+import org.apache.ignite.internal.marshaller.MarshallerColumn;
 import org.apache.ignite.lang.IgniteException;
+import org.apache.ignite.table.mapper.Mapper;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 /**
  * Client schema.
  */
+@SuppressWarnings({"rawtypes", "AssignmentOrReturnOfFieldWithMutableType", "unchecked"})
 public class ClientSchema {
     /** Schema version. Incremented on each schema modification. */
     private final int ver;
@@ -38,6 +64,15 @@ public class ClientSchema {
 
     /** Columns map by name. */
     private final Map<String, ClientColumn> map = new HashMap<>();
+
+    /** */
+    private final Map<Mapper, Marshaller> marshallers = new ConcurrentHashMap<>();
+
+    /** */
+    private final Map<Mapper, Marshaller> keyMarshallers = new ConcurrentHashMap<>();
+
+    /** */
+    private final Map<Mapper, Marshaller> valMarshallers = new ConcurrentHashMap<>();
 
     /**
      * Constructor.
@@ -117,5 +152,101 @@ public class ClientSchema {
      */
     public int keyColumnCount() {
         return keyColumnCount;
+    }
+
+    public <T> Marshaller getMarshaller(Mapper mapper) {
+        return marshallers.computeIfAbsent(mapper, m -> createMarshaller(m, TuplePart.KEY_AND_VAL));
+    }
+
+    public <T> Marshaller getKeyMarshaller(Mapper mapper) {
+        return keyMarshallers.computeIfAbsent(mapper, m -> createMarshaller(m, TuplePart.KEY));
+    }
+
+    public <T> Marshaller getValMarshaller(Mapper mapper) {
+        return valMarshallers.computeIfAbsent(mapper, m -> createMarshaller(m, TuplePart.VAL));
+    }
+
+    private Marshaller createMarshaller(Mapper mapper, TuplePart part) {
+        int colCount = columns.length;
+        int firstColIdx = 0;
+
+        if (part == TuplePart.KEY) {
+            colCount = keyColumnCount();
+        } else if (part == TuplePart.VAL) {
+            colCount = columns.length - keyColumnCount;
+            firstColIdx = keyColumnCount;
+        }
+
+        MarshallerColumn[] cols = new MarshallerColumn[colCount];
+
+        for (int i = 0; i < colCount; i++) {
+            var col = columns[i  + firstColIdx];
+
+            cols[i] = new MarshallerColumn(col.name(), mode(col.type()));
+        }
+
+        return Marshaller.createMarshaller(cols, mapper, part == TuplePart.KEY);
+    }
+
+    private static BinaryMode mode(int dataType) {
+        switch (dataType) {
+            case BOOLEAN:
+                throw new IgniteException("TODO: " + dataType);
+
+            case INT8:
+                // TODO: P_BYTE?
+                return BinaryMode.BYTE;
+
+            case INT16:
+                return BinaryMode.SHORT;
+
+            case INT32:
+                return BinaryMode.INT;
+
+            case INT64:
+                return BinaryMode.LONG;
+
+            case FLOAT:
+                return BinaryMode.FLOAT;
+
+            case DOUBLE:
+                return BinaryMode.DOUBLE;
+
+            case ClientDataType.UUID:
+                return BinaryMode.UUID;
+
+            case STRING:
+                return BinaryMode.STRING;
+
+            case BYTES:
+                return BinaryMode.BYTE_ARR;
+
+            case DECIMAL:
+                return BinaryMode.DECIMAL;
+
+            case BIGINTEGER:
+                return BinaryMode.NUMBER;
+
+            case BITMASK:
+                return BinaryMode.BITSET;
+
+            case NUMBER:
+                return BinaryMode.NUMBER;
+
+            case DATE:
+                return BinaryMode.DATE;
+
+            case TIME:
+                return BinaryMode.TIME;
+
+            case DATETIME:
+                return BinaryMode.DATETIME;
+
+            case TIMESTAMP:
+                return BinaryMode.TIMESTAMP;
+
+            default:
+                throw new IgniteException("Unknown client data type: " + dataType);
+        }
     }
 }
